requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.1
openpyxl==3.1.2
reportlab==4.0.4
markdown==3.5.1
geopy==2.4.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
